refactor: isolate openai request projection
